pyproject.toml
[tool.poetry]
name = "sketchbot"
version = "0.0.1"
description = "A Discord Bot for Chainsketch's server"
authors = ["Woid"]

[tool.poetry.dependencies]
python = "^3.10"
disnake = "^2.4.0"
odmantic = "^0.3.5"
PyYAML = "^6.0"
dnspython = "^2.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
